+++ src/Parser/Druid/Balance/Modules/Features/L90Talents.js
@@ -36,106 +36,106 @@ class L90Talents extends Analyzer {
   blessingOfElunePotential = 0;
 
   on_initialized() {
-    if (this.combatants.selected.hasTalent(SPELLS.SHOOTING_STARS_TALENT.id)){
+    if (this.combatants.selected.hasTalent(SPELLS.SHOOTING_STARS_TALENT.id)) {
       this.activeTalent = SPELLS.SHOOTING_STARS_TALENT;
     }
-    else if (this.combatants.selected.hasTalent(SPELLS.ASTRAL_COMMUNION_TALENT.id)){
+    else if (this.combatants.selected.hasTalent(SPELLS.ASTRAL_COMMUNION_TALENT.id)) {
       this.activeTalent = SPELLS.ASTRAL_COMMUNION_TALENT;
     }
-    else if (this.combatants.selected.hasTalent(SPELLS.BLESSING_OF_THE_ANCIENTS_TALENT.id)){
+    else if (this.combatants.selected.hasTalent(SPELLS.BLESSING_OF_THE_ANCIENTS_TALENT.id)) {
       this.activeTalent = SPELLS.BLESSING_OF_THE_ANCIENTS_TALENT;
     }
-    else{
+    else {
       this.active = false;
     }
   }
 
   on_byPlayer_damage(event) {
-    if ((event.ability.guid === SPELLS.MOONFIRE_BEAR.id || event.ability.guid === SPELLS.SUNFIRE.id ) && event.tick === true) {
+    if ((event.ability.guid === SPELLS.MOONFIRE_BEAR.id || event.ability.guid === SPELLS.SUNFIRE.id) && event.tick === true) {
       this.dotTicks++;
-    }       
+    }
   }
-  on_toPlayer_energize(event){
+  on_toPlayer_energize(event) {
     const spellId = event.ability.guid;
-    if (spellId !== SPELLS.LUNAR_STRIKE.id && spellId !== SPELLS.SOLAR_WRATH_MOONKIN.id){
+    if (spellId !== SPELLS.LUNAR_STRIKE.id && spellId !== SPELLS.SOLAR_WRATH_MOONKIN.id) {
       return;
     }
-    if (!this.combatants.selected.hasBuff(SPELLS.BLESSING_OF_ELUNE.id)){
+    if (!this.combatants.selected.hasBuff(SPELLS.BLESSING_OF_ELUNE.id)) {
       this.blessingOfElunePotential += this.getBonus(event.resourceChange, BLESSING_OF_ELUNE_MULTIPLIER, false);
     }
   }
-  on_changehaste(event){
-    if (this.lastHasteChangedTimestamp !== 0){
+  on_changehaste(event) {
+    if (this.lastHasteChangedTimestamp !== 0) {
       this.totalHaste += (event.timestamp - this.lastHasteChangedTimestamp) * event.oldHaste;
     }
     this.lastHasteChangedTimestamp = event.timestamp;
   }
 
-  get averageHaste(){
+  get averageHaste() {
     return this.totalHaste / this.owner.fightDuration;
   }
 
-  getGenerated(spellId){
+  getGenerated(spellId) {
     const generator = this.astralPowerTracker.buildersObj[spellId] || 0;
-    if (generator){
+    if (generator) {
       return generator.generated;
     }
     return 0;
   }
 
-  getBonus(value, increase, eluneActive){
-    if (eluneActive){
+  getBonus(value, increase, eluneActive) {
+    if (eluneActive) {
       return value - (value / (1 + increase));
     }
     return value * increase;
   }
 
-  perMinute(value){
+  perMinute(value) {
     return value / this.owner.fightDuration * 1000 * 60;
   }
 
-  get shootingStarsValue(){ 
+  get shootingStarsValue() { 
     return this.dotTicks * SHOOTING_STARS_MULTIPLIER;
   }
 
-  get astralCommunionValue(){
+  get astralCommunionValue() {
     return (1 + Math.floor(this.owner.fightDuration / 1000 / ASTRAL_COMMUNION_COOLDOWN)) * ASTRAL_COMMUNION_VALUE;
   }
 
-  get BlessingOfAnsheValue(){
+  get BlessingOfAnsheValue() {
     return (1 + Math.floor(this.owner.fightDuration / 1000 / (BLESSING_OF_ANSHE_COOLDOWN / (1 + this.averageHaste)))) * BLESSING_OF_ANSHE_VALUE;
   }
 
-  get BlessingOfEluneValue(){
+  get BlessingOfEluneValue() {
     return this.blessingOfElunePotential + this.getGenerated(SPELLS.BLESSING_OF_ELUNE.id);
   }
 
-  get actualValue(){
-    if (this.activeTalent.id === SPELLS.SHOOTING_STARS_TALENT.id){
+  get actualValue() {
+    if (this.activeTalent.id === SPELLS.SHOOTING_STARS_TALENT.id) {
       return this.getGenerated(SPELLS.SHOOTING_STARS.id);
     }
-    else if (this.activeTalent.id === SPELLS.ASTRAL_COMMUNION_TALENT.id){
+    else if (this.activeTalent.id === SPELLS.ASTRAL_COMMUNION_TALENT.id) {
       return this.getGenerated(SPELLS.ASTRAL_COMMUNION_TALENT.id);
     }
-    else if (this.activeTalent.id === SPELLS.BLESSING_OF_THE_ANCIENTS_TALENT.id){
+    else if (this.activeTalent.id === SPELLS.BLESSING_OF_THE_ANCIENTS_TALENT.id) {
       return this.getGenerated(SPELLS.BLESSING_OF_ANSHE.id) + this.getGenerated(SPELLS.BLESSING_OF_ELUNE.id);
     }
   }
 
-  get percentOfPotential(){
+  get percentOfPotential() {
     const maxPotential = Math.max(this.shootingStarsValue, this.astralCommunionValue, this.BlessingOfAnsheValue, this.BlessingOfEluneValue);
-    if (this.activeTalent.id === SPELLS.SHOOTING_STARS_TALENT.id){
+    if (this.activeTalent.id === SPELLS.SHOOTING_STARS_TALENT.id) {
       return this.shootingStarsValue / maxPotential;
     }
-    else if (this.activeTalent.id === SPELLS.ASTRAL_COMMUNION_TALENT.id){
+    else if (this.activeTalent.id === SPELLS.ASTRAL_COMMUNION_TALENT.id) {
       return this.astralCommunionValue / maxPotential;
     }
-    else if (this.activeTalent.id === SPELLS.BLESSING_OF_THE_ANCIENTS_TALENT.id){
+    else if (this.activeTalent.id === SPELLS.BLESSING_OF_THE_ANCIENTS_TALENT.id) {
       return Math.max(this.actualValue, this.BlessingOfEluneValue, this.BlessingOfAnsheValue) / maxPotential;
     }
   }
 
-  get percentOfPotentialBotA(){
+  get percentOfPotentialBotA() {
     return this.actualValue / Math.max(this.actualValue, this.BlessingOfEluneValue, this.BlessingOfAnsheValue);
   }
 
@@ -164,11 +164,11 @@ class L90Talents extends Analyzer {
 
   suggestions(when) {
     when(this.suggestionThresholds).addSuggestion((suggest, actual, recommended) => {
-        return suggest(<Wrapper><SpellLink id={this.activeTalent.id} /> only generated {formatPercentage(actual)}% of the Astral Power that another talent in the row would have generated. Consider using one of the other available choices. See the statistic box below for more information.</Wrapper>)
-          .icon(this.activeTalent.icon)
-          .actual(`${formatPercentage(actual)}% of another talent's generation.`)
-          .recommended(`>${formatPercentage(recommended)}% is recommended`);
-      });
+      return suggest(<Wrapper><SpellLink id={this.activeTalent.id} /> only generated {formatPercentage(actual)}% of the Astral Power that another talent in the row would have generated. Consider using one of the other available choices. See the statistic box below for more information.</Wrapper>)
+        .icon(this.activeTalent.icon)
+        .actual(`${formatPercentage(actual)}% of another talent's generation.`)
+        .recommended(`>${formatPercentage(recommended)}% is recommended`);
+    });
     if (this.activeTalent.id === SPELLS.BLESSING_OF_THE_ANCIENTS_TALENT.id){
       when(this.suggestionThresholdsBotA).addSuggestion((suggest, actual, recommended) => {
         return suggest(<Wrapper><SpellLink id={this.activeTalent.id} /> only generated {formatPercentage(actual)}% of the Astral Power that it could have. Make sure you are using the correct blessing. See the statistic box below for more information.</Wrapper>)
